# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.messageIn:
    properties:
      attachment_name:
        type: string
      attachment_type:
        type: string
      attachment_url:
        type: string
      content:
        type: string
      user_id:
        type: integer
    type: object
...
  handlers.telegramSend:
    properties:
      chat_id:
        type: string
      text:
        type: string
    type: object
...
  models.TelegramMessage:
    properties:
      chat_id:
        type: string
      created_at:
        type: string
      direction:
        description: in / out
        type: string
      file_id:
        description: Telegram 文件 ID
        type: string
      file_type:
        description: photo/document/video 等
        type: string
      id:
        type: integer
      is_command:
        description: 是否为 Bot 命令
        type: boolean
      text:
        type: string
      username:
        description: 发送人用户名（收件时有效）
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: 多USB 4G Modem 管理系统 API
  title: SimNexus API
  version: "1.0"
paths:
...
  /api/v1/support/conversations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: 获取所有会话列表（客服）
      tags:
      - 客服
  /api/v1/support/files/{filename}:
    get:
      parameters:
      - description: 文件名
        in: path
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 下载客服附件
      tags:
      - 客服
  /api/v1/support/messages:
    get:
      parameters:
      - description: 用户ID（客服用）
        in: query
        name: user_id
        type: integer
      - description: 增量拉取起始ID
        in: query
        name: since_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: 获取客服消息记录
      tags:
      - 客服
    post:
      consumes:
      - application/json
      parameters:
      - description: 消息内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.messageIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 发送客服消息
      tags:
      - 客服
  /api/v1/support/messages/read:
    post:
      parameters:
      - description: 用户ID（客服用）
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 标记消息已读
      tags:
      - 客服
  /api/v1/support/unread:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取未读消息数量
      tags:
      - 客服
  /api/v1/support/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 附件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 上传客服附件
      tags:
      - 客服
  /api/v1/telegram/config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取Bot配置状态
      tags:
      - Telegram
  /api/v1/telegram/file/{file_id}:
    get:
      parameters:
      - description: Telegram file_id
        in: path
        name: file_id
        required: true
        type: string
      - description: JWT令牌
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 代理下载Telegram文件
      tags:
      - Telegram
  /api/v1/telegram/messages:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 清空Telegram消息记录
      tags:
      - Telegram
    get:
      parameters:
      - description: 偏移量
        in: query
        name: skip
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TelegramMessage'
            type: array
      security:
      - BearerAuth: []
      summary: 获取Telegram消息记录
      tags:
      - Telegram
  /api/v1/telegram/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: 消息内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.telegramSend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 发送Telegram文字消息
      tags:
      - Telegram
  /api/v1/telegram/send-file:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 说明文字
        in: formData
        name: caption
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 向Telegram发送图片或文件
      tags:
      - Telegram
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"